ルーティング名称修正andコントローラーメソッド修正
