+++ laravel/resources/views/temporary.blade.php
@@ -5,7 +5,7 @@
    <div class="main">
        <h3>臨時出荷指示出力　（在庫ピンポイント指定）</h3>
     <div class="main2">
-        <form action="{{url('/temporaries')}}" method="GET">
+        <form action="{{url('/shipment/temporaries')}}" method="GET">
             <p><label for="item_code">アイテムコードを入力して下さい。
                 <input type="text" name="item_code" value="{{ $item_code ?? null }}">
             </label></p>
